Update jar_loader print_stream to work with JRuby 1.7.4
As mentioned here:
https://github.com/jenkinsci/jenkins.rb/issues/86#issuecomment-20467948
The behavior of printstream changed in JRuby 1.7. It does not accept strings. Instead of dealing with the alternative, I have just commented out the line that adds the date to the log file. I have also commented out the next line which messes up the Rails console. Sorry this is not a complete patch, just doing this now to get our project working.